Ref: DNRY for int and esm tests generation.
