+++ tools/tsconfig-base.ts
@@ -1,0 +1,4 @@
+export const getTSConfigBase = () => {
+  const nodeVersion = process.versions.node.split(".").shift();
+  return nodeVersion === "15" ? "14" : nodeVersion;
+};
